part3/requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
pandas>=2.1.0 